Silence compiler warning (negative integer implicitly converted to unsigned type)

+++ gnubg/lib/cache.c
@@ -234,8 +234,8 @@ CacheFlush(const evalCache * pc)
 {
     unsigned int k;
     for (k = 0; k < pc->size / 2; ++k) {
-        pc->entries[k].nd_primary.key.data[0] = -1;
-        pc->entries[k].nd_secondary.key.data[0] = -1;
+        pc->entries[k].nd_primary.key.data[0] = (unsigned int) -1;
+        pc->entries[k].nd_secondary.key.data[0] = (unsigned int) -1;
 #if USE_MULTITHREAD
         pc->entries[k].lock = 0;
 #endif
